resolving netlify build
